Adding of further tests. Adding of Test/Sample Data Adding of Console Application Adding Error Handlers Fixed bugs

+++ src/main/java/au/com/fx/converter/service/FxConversionService.java
@@ -1,9 +1,11 @@
 package au.com.fx.converter.service;
 
+import java.math.BigDecimal;
+
 /**
  * Created by senthurshanmugalingm on 6/05/2017.
  */
 public interface FxConversionService {
 
-    Double convert(String baseCurrency, String termCurrency, Double amount);
+    BigDecimal convert(String baseCurrency, String termCurrency, Double amount);
 }
